Update verify token for auths controllers

+++ src/controllers/companyController.ts
@@ -46,31 +46,43 @@ export const getCompany = async(req: Request, res: Response, next: NextFunction)
 }
 
 export const updateCompany = async(req: Request, res: Response, next: NextFunction) => {
-  try {
-    const {password, ...bodyDetails } = req.body;
+  
+  if(req.cookies.cookies.id === req.params.companyId && req.cookies.cookies.isAdmin){
 
-    const company: CompanyType | null= await Company.findByIdAndUpdate(req.params.companyId, {$set: bodyDetails}, {new: true})
-    
-    if(!company) return res.status(400).send("Record does not exist")
-    
-    const { password: companyPassword, ...otherDetails } = company._doc;
-    
-    res.status(200).json(otherDetails);
-  } catch (err) {
-    next(err)
+    try {
+      const {password, ...bodyDetails } = req.body;
+  
+      const company: CompanyType | null= await Company.findByIdAndUpdate(req.params.companyId, {$set: bodyDetails}, {new: true})
+      
+      if(!company) return res.status(400).send("Record does not exist")
+      
+      const { password: companyPassword, ...otherDetails } = company._doc;
+      
+      res.status(200).json(otherDetails);
+    } catch (err) {
+      next(err)
+    }
+  }else {
+    res.status(401).send("You are not authorised to make changes")
   }
 }
 
 export const deleteCompany = async(req: Request, res: Response, next: NextFunction) => {
-  try {
-    
-    await Company.findByIdAndDelete(req.params.companyId);
-    const offices = await Officer.find({companyId: req.params.companyId})
-    offices.forEach(item => item.deleteOne())
-    
-    res.status(200).send("Company deleted successfully");
+  if(req.cookies.cookies.id === req.params.companyId && req.cookies.cookies.isAdmin){
 
-  } catch (err) {
-    next(err)
+    try {
+      
+      await Company.findByIdAndDelete(req.params.companyId);
+      const offices = await Officer.find({companyId: req.params.companyId})
+  
+      offices.forEach(item => item.deleteOne())
+      
+      res.status(200).send("Company deleted successfully");
+  
+    } catch (err) {
+      next(err)
+    }
+  }else {
+    res.status(401).send("You are not authorised to perform this action")
   }
 }
